+++ Convertor/ImageBase/Header.cs
@@ -2,7 +2,7 @@
 {
     public abstract class Header
     {
-        public abstract int Height { get; set; }
-        public abstract int Width { get; set; }
+        public abstract short Height { get; set; }
+        public abstract short Width { get; set; }
     }
 }
